Version2211

+++ src/MolecularMechanics/ConjugateGradient.c
@@ -30,7 +30,7 @@ DEALINGS IN THE SOFTWARE.
 #include "ConjugateGradient.h"
 
 
-static float maxarg1,maxarg2;
+static gdouble maxarg1,maxarg2;
 #define FMAX(a,b) (maxarg1=(a),maxarg2=(b),(maxarg1) > (maxarg2) ?\
         (maxarg1) : (maxarg2))
